# api/docs/swagger.yaml
basePath: /
definitions:
...
  apidocs.ErrorBody:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
...
  apidocs.InputSwitchRequest:
    properties:
      priority:
        type: integer
    type: object
...
  apidocs.StreamActionData:
    properties:
      data:
        $ref: '#/definitions/apidocs.StreamActionInner'
    type: object
  apidocs.StreamActionInner:
    properties:
      status:
        type: string
    type: object
  apidocs.StreamData:
    properties:
      data:
        $ref: '#/definitions/apidocs.StreamResponse'
    type: object
  apidocs.StreamList:
    properties:
      data:
        items:
          $ref: '#/definitions/apidocs.StreamResponse'
        type: array
      total:
        type: integer
    type: object
  apidocs.StreamPutRequest:
    properties:
      description:
        type: string
      disabled:
        type: boolean
      dvr:
        $ref: '#/definitions/domain.StreamDVRConfig'
      inputs:
        items:
          $ref: '#/definitions/domain.Input'
        type: array
      name:
        type: string
      protocols:
        $ref: '#/definitions/domain.OutputProtocols'
      push:
        items:
          $ref: '#/definitions/domain.PushDestination'
        type: array
      stream_key:
        type: string
      tags:
        items:
          type: string
        type: array
      thumbnail:
        $ref: '#/definitions/domain.ThumbnailConfig'
      transcoder:
        $ref: '#/definitions/domain.TranscoderConfig'
      watermark:
        $ref: '#/definitions/domain.WatermarkConfig'
    type: object
  apidocs.StreamResponse:
    properties:
      code:
        description: Code is the unique key chosen by the user ([a-zA-Z0-9_]).
        type: string
      description:
        type: string
      disabled:
        description: Disabled when true excludes the stream from server bootstrap
          and rejects pipeline Start.
        type: boolean
      dvr:
        allOf:
        - $ref: '#/definitions/domain.StreamDVRConfig'
        description: |-
          DVR overrides the global DVR settings for this specific stream.
          If nil, the global config is used (when DVR is enabled globally).
      inputs:
        description: |-
          Inputs are the available ingest sources ordered by Priority.
          The Stream Manager monitors health and switches between them on failure.
        items:
          $ref: '#/definitions/domain.Input'
        type: array
      name:
        type: string
      protocols:
        allOf:
        - $ref: '#/definitions/domain.OutputProtocols'
        description: |-
          Protocols defines which delivery protocols are opened for this stream.
          The server opens a listener/packager for each enabled protocol.
          Protocol-level config (ports, segment duration, CDN URL) lives in server config.
      push:
        description: |-
          Push is the list of external destinations the server actively pushes to.
          Each entry defines one push target (YouTube, Facebook, Twitch, CDN relay, etc.).
        items:
          $ref: '#/definitions/domain.PushDestination'
        type: array
      runtime:
        $ref: '#/definitions/manager.RuntimeStatus'
      stream_key:
        description: StreamKey is used to authenticate RTMP/SRT push ingest.
        type: string
      tags:
        items:
          type: string
        type: array
      thumbnail:
        allOf:
        - $ref: '#/definitions/domain.ThumbnailConfig'
        description: Thumbnail controls periodic screenshot generation for preview
          images.
      transcoder:
        allOf:
        - $ref: '#/definitions/domain.TranscoderConfig'
        description: |-
          Transcoder controls encoding/decoding settings.
          nil means no transcoding for this stream.
      watermark:
        allOf:
        - $ref: '#/definitions/domain.WatermarkConfig'
        description: Watermark is an optional text or image overlay applied before
          encoding.
    type: object
...
  domain.DecoderConfig:
    properties:
      name:
        description: |-
          Name is the FFmpeg decoder name.
          "" = let FFmpeg choose automatically.
          Examples: "h264_cuvid", "h264_qsv".
        type: string
    type: object
...
  domain.Input:
    properties:
      headers:
        additionalProperties:
          type: string
        description: |-
          Headers are arbitrary HTTP headers sent with every request for HTTP/HLS inputs.
          Common uses:
            "Authorization": "Bearer <token>"
            "Authorization": "Basic <base64(user:pass)>"
            "X-Custom-Token": "secret"
        type: object
      net:
        allOf:
        - $ref: '#/definitions/domain.InputNetConfig'
        description: Net controls reconnect and timeout behaviour.
      params:
        additionalProperties:
          type: string
        description: |-
          Params are extra URL query parameters merged into the source URL before
          connecting. Used for protocols that carry credentials or options in the
          query string (SRT ?passphrase=, S3 ?access_key= / ?secret_key=, etc.).
            "passphrase": "my-srt-passphrase"
            "access_key": "AKID..."   (S3)
            "secret_key": "wJal..."   (S3)
        type: object
      priority:
        description: |-
          Priority determines failover order. Lower value = higher priority.
          The Stream Manager always prefers the lowest-priority alive input.
        type: integer
      url:
        description: URL is the source endpoint. See the package doc for supported
          formats.
        type: string
    type: object
...
  domain.OutputProtocols:
    properties:
      dash:
        description: |-
          DASH enables MPEG-DASH packaging over HTTP.
          Required for Widevine/PlayReady DRM.
        type: boolean
      hls:
        description: |-
          HLS enables Apple HTTP Live Streaming (m3u8 + segments over HTTP).
          Compatible with browsers, iOS, Android, Smart TVs.
        type: boolean
      rtmp:
        description: RTMP opens an RTMP publish endpoint for legacy players/CDNs.
        type: boolean
      rtsp:
        description: RTSP opens an RTSP listener for pull clients (VLC, broadcast
          tools).
        type: boolean
      srt:
        description: SRT opens an SRT listener port for contribution-quality pull.
        type: boolean
    type: object
...
  domain.PushDestination:
    properties:
      comment:
        description: Comment is a human-readable note for this destination.
        type: string
      enabled:
        description: Enabled controls whether this destination is active.
        type: boolean
      limit:
        description: Limit is the maximum number of retry attempts. 0 = unlimited.
        type: integer
      retry_timeout_sec:
        description: RetryTimeoutSec is the delay between retry attempts in seconds.
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/domain.PushStatus'
        description: |-
          Status is a runtime-only field updated by the publisher.
          Not persisted to storage.
      timeout_sec:
        description: TimeoutSec is the connection/write timeout in seconds.
        type: integer
      url:
        description: |-
          URL is the destination ingest endpoint.
          Supported schemes:
            rtmp://  — plain TCP, default port 1935 (e.g. rtmp://a.rtmp.youtube.com/live2/{key})
            rtmps:// — TLS-wrapped RTMP, default port 443 (e.g. rtmps://live-api-s.facebook.com:443/rtmp/{key})
        type: string
    type: object
  domain.PushStatus:
    enum:
    - idle
    - connecting
    - active
    - retrying
    - failed
    - disabled
    type: string
    x-enum-varnames:
    - PushStatusIdle
    - PushStatusConnecting
    - PushStatusActive
    - PushStatusRetrying
    - PushStatusFailed
    - PushStatusDisabled
...
  domain.StreamDVRConfig:
    properties:
      enabled:
        type: boolean
      max_size_gb:
        description: |-
          MaxSizeGB caps total disk usage. Oldest segments pruned when exceeded.
          0 = no limit.
        type: number
      retention_sec:
        description: |-
          RetentionSec is the retention window in seconds.
          0 = keep forever.
        type: integer
      segment_duration:
        description: |-
          SegmentDuration overrides the global segment length in seconds.
          0 = use default (4s).
        type: integer
      storage_path:
        description: |-
          StoragePath overrides the default DVR root directory for this stream.
          "" = use "./dvr/{streamCode}".
        type: string
    type: object
...
  domain.ThumbnailConfig:
    properties:
      enabled:
        type: boolean
      height:
        type: integer
      interval_sec:
        description: IntervalSec generates one thumbnail every N seconds.
        type: integer
      output_dir:
        description: |-
          OutputDir is relative to the publisher HLS directory.
          E.g. "thumbnails" → written to {hls_dir}/{stream_code}/thumbnails/thumb.jpg
        type: string
      quality:
        description: 'Quality is the JPEG quality (1–31, lower = better). Default:
          5.'
        type: integer
      width:
        description: |-
          Width and Height of the output thumbnail in pixels.
          0 = match source resolution.
        type: integer
    type: object
  domain.TranscoderConfig:
    properties:
      audio:
        $ref: '#/definitions/domain.AudioTranscodeConfig'
      decoder:
        $ref: '#/definitions/domain.DecoderConfig'
      extra_args:
        description: |-
          ExtraArgs are raw FFmpeg arguments appended after the generated command.
          Use with caution — may conflict with generated arguments.
        items:
          type: string
        type: array
      global:
        $ref: '#/definitions/domain.TranscoderGlobalConfig'
      video:
        $ref: '#/definitions/domain.VideoTranscodeConfig'
    type: object
...
  domain.WatermarkConfig:
    properties:
      asset_id:
        description: |-
          AssetID, when set, references a WatermarkAsset uploaded via the
          /watermarks API. Coordinator resolves it to an on-disk file path
          before passing the config to the transcoder. Takes precedence over
          ImagePath when both are set.
        type: string
      enabled:
        type: boolean
      font_color:
        description: FontColor in FFmpeg color syntax. E.g. "white", "#FFFFFF", "white@0.8".
        type: string
      font_file:
        description: |-
          FontFile is the path to a .ttf/.otf font file.
          "" = FFmpeg default font.
        type: string
      font_size:
        description: 'FontSize in pixels. Default: 24.'
        type: integer
      image_path:
        description: |-
          ImagePath is the absolute path to a watermark image (PNG with alpha
          recommended). Use this for assets pre-staged on the host outside the
          /watermarks library. Mutually exclusive with AssetID.
        type: string
      offset_x:
        description: |-
          OffsetX and OffsetY are pixel offsets from the chosen position edge.
          Ignored when Position == custom.
        type: integer
      offset_y:
        type: integer
      opacity:
        description: 'Opacity controls transparency: 0.0 = fully transparent, 1.0
          = fully opaque.'
        type: number
      position:
        allOf:
        - $ref: '#/definitions/domain.WatermarkPosition'
        description: |-
          Position selects how the (X, Y) of the overlay are computed:
            - top_left/top_right/bottom_left/bottom_right/center: convenience
              anchors. OffsetX/Y act as inward padding from the chosen edge
              (Center ignores offsets).
            - custom: X / Y below are used as raw FFmpeg expressions —
              pixel ints ("100"), expressions ("main_w-overlay_w-50"), or
              time-aware fades ("if(gt(t,5),10,-100)") all work.
      text:
        description: |-
          Text is the string to render. Supports strftime directives for live timestamps.
          E.g. "LIVE %{localtime:%H:%M:%S}"
        type: string
      type:
        $ref: '#/definitions/domain.WatermarkType'
      x:
        description: |-
          X / Y are raw FFmpeg coordinate expressions used only when Position
          == custom. Empty string defaults to "0". The exact variables exposed
          depend on the filter:
            - drawtext (text watermark): w/h = frame size, tw/th = text size
            - overlay (image watermark):  W/H = main video size, w/h = overlay size
        type: string
      "y":
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for live stream configuration, ingest/publish lifecycle, DVR
    recordings, and event hooks.
  title: Open Streamer API
  version: "1.0"
paths:
...
  /streams:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: List streams
      tags:
      - streams
  /streams/{code}:
    delete:
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Delete stream
      tags:
      - streams
    get:
      parameters:
      - description: Stream code (a-zA-Z0-9_)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Get stream
      tags:
      - streams
    post:
      consumes:
      - application/json
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      - description: Partial or full stream document
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.StreamPutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.StreamData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Create or partial-update stream
      tags:
      - streams
  /streams/{code}/inputs/switch:
    post:
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      - description: Target input priority
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.InputSwitchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamActionData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Manual input switch
      tags:
      - streams
...
  /streams/{code}/restart:
    post:
      parameters:
      - description: Stream code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.StreamActionData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorBody'
      summary: Restart stream pipeline
      tags:
      - streams
...
swagger: "2.0"